Crate hafas_wrap

Hafas-wrap Library

The hafas-wrap crate provides an interface for HAFAS. This is done using static request and response objects which can be parsed into json strings, requested and parsed back into objects.

Structure and usage

The main functionality comes with the Client, see the documentation for it on how to use this library.

This library also provides some hafas_profiles for common endpoints and request and response structs to simplyfy the working with HAFAS endpoints. The util module provides functions to de- and encode Base64-AES Strings using a statically set key and a md5 hashing function needed for requests.

Modules

hafas_profiles

A collection of different HAFAS profiles for different endpoints.

methods

Predefined request and response structs for HAFAS requests.

util

Misc functions

Macros

define_request

Creates a request struct with corresponding implementations

Structs

Client

The client is a simple struct that manages requests to HAFAS. I only contains a reqwest web client to make requests.

RawHafasRequest

The RawHafasRequest consists of both the request method and a generic request object that has to implement Serialize. It is used to query HAFAS in request().

Enums

RequestError

The request error is the error returned from the two request functions.