Macro hafas_wrap::define_request

macro_rules! define_request {
    (
        $method_name:expr,
        $struct_name:ident {
            $($field_name:ident: $field_type:ty,)*
        }
    ) => { ... };
}

Creates a request struct with corresponding implementations

This macro creates a struct with a From implementation into a RawHafasRequest and some needed deriviatives like Serialize, Debug and Default. It also allows non_snake_case because most of HAFAS is written in camelCase.

Altrough this macro derives Debug, that is just to simplify the creation of request objects. It it not safe to assume the default is a working request.

Example

define_request!(
    "ExampleMethodName",
    ExampleMethodRequest {
        some_value: bool,
        longer_value: String,
    }
);