Function hafas_wrap::util::decode

pub fn decode(secret: &str) -> Result<String, AESDecodeError>

Decrypt a Base64-AES encoded String.

This function uses the openssl library to decrypt a given String together with a statically set key.

Errors

FromUtf8Error when the string is not correctly utf-8 encoded.

DecryptError when the string is wrongly encrypted.

DecodeError when the string is wrongly base-64 encoded.